requirements.txt
opencv-python
numpy
tqdm
click
tkinterdnd2
Pillow